+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Change log
 
 - Inline CKEditor: Dropped the admin jQuery dependency for real.
 - Started using pyupgrade_ for the Python code.
+- Added Django 4.0a1 to the CI matrix.
 
 
 `0.93`_ (2021-09-20)