+++ CHANGELOG.rst
@@ -6,6 +6,10 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+
+`0.94`_ (2021-09-29)
+~~~~~~~~~~~~~~~~~~~~
+
 - Inline CKEditor: Dropped the admin jQuery dependency for real.
 - Started using pyupgrade_ for the Python code.
 - Added Django 4.0a1 to the CI matrix.
@@ -843,4 +847,5 @@ functionality.
 .. _0.91: https://github.com/matthiask/feincms3/compare/0.90...0.91
 .. _0.92: https://github.com/matthiask/feincms3/compare/0.91...0.92
 .. _0.93: https://github.com/matthiask/feincms3/compare/0.92...0.93
-.. _Next version: https://github.com/matthiask/feincms3/compare/0.93...main
+.. _0.94: https://github.com/matthiask/feincms3/compare/0.93...0.94
+.. _Next version: https://github.com/matthiask/feincms3/compare/0.94...main